refactor: remove render header
We simplify the coding more by removing the small render header and
making the functions in it methods of the appropriate classes. We also
move some more functions to the more appropriate code unit.
